refactor(jira): split the oversized module along two real seams
CodeRabbit flagged `jira/mod.rs` at 648 lines against the repo's 500-line rule.
Splitting it revealed that my own JQL-fallback commit had pushed `fetch.rs` from
492 to 640, so this fixes both.

- `refresh.rs` takes the refresh workflow - the staleness gate, the
  fetch-upsert-prune sequence, and the two public entry points. mod.rs keeps the
  wire DTOs and the per-row database operations those steps call. That is the
  seam CodeRabbit suggested and it is a real one: orchestration on one side,
  the operations it sequences on the other.
- `active_filter.rs` takes what counts as an active work item - the two JQL
  strings, the container-hierarchy rule, the name policy, and the
  unknown-issue-type classifier. `fetch.rs` keeps the HTTP mechanics that carry
  them. Policy on one side, transport on the other.

mod.rs 648 -> 460, fetch.rs 640 -> 478. Every production module in this
directory is now under the cap; `tests.rs` (575) is a test file and untouched.

Worth recording: mod.rs was ALREADY 606 lines on main, so this release did not
create the violation - it added 42 lines to a standing one. Same class as #863.
Fixed here rather than deferred because the release PR cannot merge with the
conversation open, and because splitting it is compiler-verified and covered by
61 jira tests.

The moved code's `super::` referred to `providers`; inside the new modules it
would mean `jira`, so those five call sites now use absolute paths.
